touch multiple files, ask if overwriting is ok, if file already exists and template is given

+++ src/lib.rs
@@ -8,6 +8,7 @@ use std::path::{Path, PathBuf};
 use dirs::home_dir;
 
 use clap::Parser;
+use dialoguer::Confirm;
 use tera::{Context, Tera};
 use thiserror::Error;
 
@@ -15,15 +16,15 @@ use thiserror::Error;
 #[clap(name = "zap", author, version, about = "touch, but with templates", long_about = None)]
 pub struct ZapCli {
     #[clap(value_parser)]
-    pub filename: String,
+    pub filenames: Vec<String>,
 
-    /// Optional template name to pre-populate the file.\n
+    /// Optional template name to pre-populate the file.
     /// Templates are sourced from ~/.config/zap/<template_name>.
-    #[clap(short = 't', long, value_name = "TEMPLATE_NAME")]
+    #[clap(short = 'T', long, value_name = "TEMPLATE_NAME", verbatim_doc_comment)]
     pub template: Option<String>,
 
     /// Optional context to use when rendering the template.
-    #[clap(short = 'c', long, value_name = "CONTEXT")]
+    #[clap(short = 'C', long, value_name = "CONTEXT")]
     pub context: Option<String>,
 
     /// Open the file with your $EDITOR
@@ -43,6 +44,8 @@ pub enum ZapError {
     TemplateNotFound(PathBuf),
     #[error("Failed to set file times: {0}")]
     SetTimesError(io::Error),
+    #[error("Dialoguer error: {0}")]
+    Dialoguer(#[from] dialoguer::Error),
 
     #[error("EDITOR environment variable not set")]
     EditorNotSet,
@@ -69,47 +72,57 @@ fn get_template_path(template_name: &str) -> Result<PathBuf, ZapError> {
 /// zap: Create a file if it doesn't exist,
 /// optionally populate it with text from a template.
 /// If the file exists, its modification and access times are updated.
-pub fn zap(filename_str: &str, template_name: Option<&str>, context_str: Option<&str>) -> Result<(), ZapError> {
-    let path = Path::new(filename_str);
+pub fn zap(filenames: &Vec<String>, template_name: Option<&str>, context_str: Option<&str>) -> Result<(), ZapError> {
 
-    // no template provided and the file exists: Just update the timestamp
-    if path.exists() {
+    for filename in filenames {
+
+        let path = Path::new(filename);
         let now = FileTime::now();
-        if template_name.is_none() {
-            return SetFileTimes(path, now, now)
-                .map_err(ZapError::SetTimesError);
-        }
-    }
 
-    let mut file = File::create(path)?;
-
-    if let Some(tmpl_name) = template_name {
-        let template_path_full = get_template_path(tmpl_name)?;
-
-        if !template_path_full.exists() {
-            return Err(ZapError::TemplateNotFound(template_path_full));
-        }
-
-        let mut tera = Tera::default();
-
-        tera.add_template_file(&template_path_full, Some(tmpl_name))?;
-
-        let mut context = Context::new();
-        if let Some(ctx) = &context_str {
-            for pair in ctx.split(',') {
-                let mut parts = pair.splitn(2, '=');
-                if let (Some(key), Some(value)) = (parts.next(), parts.next()) {
-                    context.insert(key.trim(), value.trim());
+        if path.exists() {
+            if template_name.is_none() { // If no template is provided, just update the file times
+                return SetFileTimes(path, now, now).map_err(ZapError::SetTimesError);
+            } else {
+                let confirmation = Confirm::new()
+                    .with_prompt(format!("File '{}' already exists. Do you want to overwrite it?", filename))
+                    .default(false)
+                    .interact()?;
+                if !confirmation {
+                    continue; // Skip to the next file if user doesn't agree with overwrite
                 }
             }
         }
-        let rendered = tera.render(tmpl_name, &context)?;
-        file.write_all(rendered.as_bytes())?;
+
+        let mut file = File::create(path)?;
+
+        if let Some(tmpl_name) = template_name {
+            let template_path_full = get_template_path(tmpl_name)?;
+
+            if !template_path_full.exists() {
+                return Err(ZapError::TemplateNotFound(template_path_full));
+            }
+
+            let mut tera = Tera::default();
+
+            tera.add_template_file(&template_path_full, Some(tmpl_name))?;
+
+            let mut context = Context::new();
+            if let Some(ctx) = &context_str {
+                for pair in ctx.split(',') {
+                    let mut parts = pair.splitn(2, '=');
+                    if let (Some(key), Some(value)) = (parts.next(), parts.next()) {
+                        context.insert(key.trim(), value.trim());
+                    }
+                }
+            }
+            let rendered = tera.render(tmpl_name, &context)?;
+            file.write_all(rendered.as_bytes())?;
+        }
     }
     Ok(())
 }
 
-pub fn open_in_editor(filepath: &str) -> Result<(), ZapError> {
+pub fn open_in_editor(filepaths: &Vec<String>) -> Result<(), ZapError> {
     let editor_env_var = env::var("EDITOR").map_err(|_| ZapError::EditorNotSet)?;
 
     let mut parts = editor_env_var.split_whitespace();
@@ -118,7 +131,7 @@ pub fn open_in_editor(filepath: &str) -> Result<(), ZapError> {
 
     let mut cmd = Command::new(editor_executable);
     cmd.args(parts);
-    cmd.arg(filepath);
+    cmd.args(filepaths);
 
     match cmd.status() {
         Ok(status) => {
